requirements.txt
pytorch
torchvision
opencv-python 